locations works
